First Working Commit
This is the first working commit of the LRU Cache

+++ src/nisran/cache/LRUCacheNode.java
@@ -1,18 +1,17 @@
 package nisran.cache;
 
-import java.math.BigInteger;
 
 public class LRUCacheNode<V extends Object> {
 	
 	/* Key Value Store */
-	BigInteger key;
+	Integer key;
 	V value;
 	
 	/* Bi-directional List 	 */
 	LRUCacheNode<V> previous;
 	LRUCacheNode<V> next;
 	
-	public LRUCacheNode(BigInteger key, V value) {
+	public LRUCacheNode(Integer key, V value) {
 		this.key = key;
 		this.value = value;
 	}
